requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.3.1
django-filter==24.2
django-simple-history==3.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
sqlparse==0.5.0
tzdata==2024.1
whitenoise==6.6.0
